+++ drivers/pci/pci-sysfs.c
@@ -711,7 +711,7 @@ pci_mmap_resource(struct kobject *kobj, struct bin_attribute *attr,
 			"at page 0x%08lx on %s BAR %d (start 0x%16Lx, size 0x%16Lx)\n",
 			current->comm, vma->vm_end-vma->vm_start, vma->vm_pgoff,
 			pci_name(pdev), i,
-			pci_resource_start(pdev, i), pci_resource_len(pdev, i));
+			(long long unsigned int)pci_resource_start(pdev, i), (long long unsigned int)pci_resource_len(pdev, i));
 		return -EINVAL;
 	}
